opl: Use uint64_t for microsecond times.
Previous revamping of the OPL library changed delay times from ms
to us, and switched to 64-bit times for storage for them. However
this type change was botched and some functions were still using
32-bit arguments. Change to 64-bit types.

Thanks Quasar for reporting this. This fixes #438.

+++ opl/opl.h
@@ -110,7 +110,7 @@ void OPL_InitRegisters(void);
 // Set a timer callback.  After the specified number of microseconds
 // have elapsed, the callback will be invoked.
 
-void OPL_SetCallback(unsigned int us, opl_callback_t callback, void *data);
+void OPL_SetCallback(uint64_t us, opl_callback_t callback, void *data);
 
 // Adjust callback times by the specified factor. For example, a value of
 // 0.5 will halve all remaining times.
@@ -132,7 +132,7 @@ void OPL_Unlock(void);
 
 // Block until the specified number of microseconds have elapsed.
 
-void OPL_Delay(unsigned int us);
+void OPL_Delay(uint64_t us);
 
 // Pause the OPL callbacks.
